Authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -64,7 +64,13 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        // $this->authorize("update", $question);
+        // if(\Gate::allows('update-question', $question)){
+        //     return view("questions.edit", compact('question'));
+        // }
+
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied by MR");
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -77,6 +83,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied by MR");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', "Your question has been updated.");
@@ -90,6 +101,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Access denied by MR");
+        }
+
         // $this->authorize("delete", $question);
 
         $question->delete();
